Rename cli helper to progress_helper and extract non-progress methods to base cli class
